fixes bug unable to import statement. We are unable to delete bugged folder
