+++ client/package.json
@@ -11,7 +11,7 @@
     "mkdirp": "^1.0.4",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.3",
+    "react-scripts": "^3.4.4",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^1.2.1"
   },
